+++ Patches/HUDManager_Patching.cs
@@ -9,6 +9,8 @@ internal static class HUDManager_Patching
     [HarmonyPrefix]
     private static bool CanPlayerScan_Prefix(ref bool __result)
     {
+        if (!PingScan.DoPatch) return true;
+        
         __result = false;
         return false;
     }